requirements.txt
openai==0.27.2
python-decouple==3.8
selenium==4.8.2
termcolor==2.2.0
